tests/integration/requirements.txt
pytest>=7.0
pyyaml>=6.0
grpcio>=1.60.0
protobuf>=4.25.0
